Extract link search on fields into own class.
I moved all code responsible for link search in fields into its own class to split different task into different units. Also I was able to reduce the complexity of the unittests because now I don't need the platform anymore (when instancing a LinkOnFieldSeeker instance).

Since the module does not only contain the accumulator from now on I tried to give it a better fitting name as well.
